Implement reverse lookup output (IPv6, subset)
Add reverse lookup with "correct" output formatting for IPv6
in hexadecimal colon format. While the most common, this is only
a subset of all correct IPv6 formats.

Regexp::Common in version 20130301 has support for a correct IPv6
regexp, but I did not want to add an external dependency. This is
a dirty debug helper; elevating it from that to general purpose tool
will take a lot more :)